Extract verticalMultiline impl from Router
This should help with the ever-growing compilation times.
