+++ core/src/main/scala/org/scalafmt/internal/Router.scala
@@ -423,125 +423,11 @@ class Router(formatOps: FormatOps) {
 
       // Parameter opening for one parameter group. This format works
       // on the WHOLE defnSite (via policies)
-      case FormatToken(open @ (LeftParen() | LeftBracket()), r, between)
+      case ft @ FormatToken((LeftParen() | LeftBracket()), _, _)
           if style.verticalMultilineAtDefinitionSite &&
             isClassOrDef(leftOwner) =>
 
-        val close = matchingParentheses(hash(open))
-        val indentParam = Num(style.continuationIndent.defnSite)
-        val indentSep = Num((indentParam.n - 2).max(0))
-        val isBracket = open.is[LeftBracket]
-
-        @tailrec
-        def loop(token: Token): Option[Token] = {
-          leftTok2tok(matchingParentheses(hash(token))) match {
-            case FormatToken(RightParenOrBracket(), l @ LeftParen(), _) =>
-              loop(l)
-            // This case only applies to classes
-            case f @ FormatToken(RightBracket(), mod, _)
-              if mod.is[Modifier] =>
-              loop(next(f).right)
-            case FormatToken(r @ RightParenOrBracket(), _, _) => Some(r)
-            case _ => None
-          }
-        }
-        // find the last param on the defn so that we can apply our `policy`
-        // till the end.
-        val lastParen = loop(open).get
-
-        val mixedParams = {
-          leftOwner match {
-            case cls: meta.Defn.Class =>
-              cls.tparams.nonEmpty && cls.ctor.paramss.nonEmpty
-            case _ => false
-          }
-        }
-
-        val shouldNotDangle =
-          leftOwner.is[meta.Ctor.Primary] ||
-          leftOwner.is[meta.Defn.Class] ||
-          leftOwner.is[meta.Defn.Trait]
-
-        // Since classes and defs aren't the same (see below), we need to
-        // create two (2) OneArgOneLineSplit when dealing with classes. One
-        // deals with the type params and the other with the value params.
-        val oneLinePerArg = {
-          val base = OneArgOneLineSplit(open)
-          if (mixedParams) {
-            val afterTypes = leftTok2tok(matchingParentheses(hash(open)))
-            // Try to find the first paren. If found, then we are dealing with
-            // a class with type AND value params. Otherwise it is a class with
-            // just type params.
-            findFirst(afterTypes, lastParen)(t => t.left.is[LeftParen])
-              .fold(base)(t => base.merge(OneArgOneLineSplit(t.left)))
-          }
-          else base
-        }
-
-        // DESNOTE(2017-03-28, pjrt) Classes and defs aren't the same.
-        // For defs, type params and value param have the same `owners`. However
-        // this is not the case for classes. Type params have the class itself
-        // as the owner, but value params have the Ctor as the owner, so a
-        // simple check isn't enough. Instead we check against the owner of the
-        // `lastParen` as well, which will be the same as the value param's
-        // owner.
-        val valueParamsOwner = owners(lastParen)
-        @inline def ownerCheck(rp: Token): Boolean = {
-          val owner = owners(rp)
-          owner == leftOwner || owner == valueParamsOwner
-        }
-
-        val paramGroupSplitter: PartialFunction[Decision, Decision] = {
-          // If this is a class, then don't dangle the last paren
-          case Decision(t @ FormatToken(_, rp @ RightParenOrBracket(), _), _)
-              if shouldNotDangle && rp == lastParen =>
-            val split = Split(NoSplit, 0)
-            Decision(t, Seq(split))
-          // Indent seperators `)(` and `](` by `indentSep`
-          case Decision(t @ FormatToken(_, rp @ RightParenOrBracket(), _), _)
-              if ownerCheck(rp) =>
-            val split = Split(Newline, 0).withIndent(indentSep, rp, Left)
-            Decision(t, Seq(split))
-          // Do NOT Newline the first param after the split, unless we have a
-          // mixed-params case with a Ctor modifier.
-          // `] private (`
-          case Decision(t @ FormatToken(open2 @ LeftParen(), _, _), _) =>
-            val close2 = matchingParentheses(hash(open2))
-            val prevT = prev(t).left
-            val mod =
-              if (mixedParams && owners(prevT).is[CtorModifier]) Newline
-              else NoSplit
-            Decision(t,
-                     Seq(
-                       Split(mod, 0)
-                         .withIndent(indentParam, close2, Right)
-                     ))
-        }
-
-        // Our policy is a combination of OneArgLineSplit and a custom splitter
-        // for parameter groups.
-        val policy = oneLinePerArg.merge(paramGroupSplitter, lastParen.end)
-
-        val firstIndent =
-          if (r.is[RightParen]) // An empty param group
-            indentSep
-          else
-            indentParam
-
-        val singleLineSplit =
-          if (isBracket)
-            Split(NoSplit, 0) // If we can fit the type params, make it so
-              .withPolicy(SingleLineBlock(close))
-          else
-            Split(NoSplit, 0) // If we can fit all in one block, make it so
-              .withPolicy(SingleLineBlock(lastParen))
-
-        Seq(
-          singleLineSplit,
-          Split(Newline, 1) // Otherwise split vertically
-            .withIndent(firstIndent, close, Right)
-            .withPolicy(policy)
-        )
+          verticalMultiline(leftOwner, ft)(style)
 
       // Term.Apply and friends
       case FormatToken(LeftParen() | LeftBracket(), right, between)
